Added Deploy stage to Jenkins

+++ Jenkinsfile
@@ -14,6 +14,12 @@ node {
         sh 'docker build -t app --no-cache .'
         sh 'docker tag app localhost:5000/app'
         sh 'docker push localhost:5000/app'
+      }
+    }
+    stage('Deploy'){
+      if(env.BRANCH_NAME == 'master'){
+        sh 'docker pull localhost:5000/app'
+        sh 'docker run -d -p 8090:8090 --name app localhost:5000/app:latest'
         sh 'docker rmi -f app localhost:5000/app'
       }
     }
